First commit after a long time

+++ CL_flightData.php
@@ -173,6 +173,28 @@ var $maxPointNum=1000;
 
 	}
 
+	function flightIsViewableByUser($userID) {
+		if ($this->belongsToUser($userID)) return 1;
+
+		global $db;
+
+		$queryFriends="SELECT * FROM leonardo_friends WHERE userID=".$this->userID;
+		$resFriends= $db->sql_query($queryFriends);
+
+		if($resFriends<= 0){
+			echo("<H3> Error in query: $queryFriends</H3>\n");
+			return 0;
+		}
+		$friendsIDs=array();
+		while($rowFriends= $db->sql_fetchrow($resFriends) ){
+			$friendsIDs[$rowFriends['friendUserID']]=$rowFriends['friendUserID'];
+		}
+
+		//echo $queryFriends. " $userID #";
+		if ($friendsIDs[$userID]) return 1;
+		else return 0;
+	}
+
 	function checkGliderBrand($gliderBrand='') {
 
 		if (! $this->gliderBrandID ) {
@@ -281,7 +303,7 @@ var $maxPointNum=1000;
 		$lastPoint->setLon($this->lastLon);
 		$lastPoint->gpsTime=$this->lastPointTM;
 
-		list($lastName,$firstName,$pilotCountry,$Sex,$Birthdate,$CIVL_ID)=getPilotInfo($this->userID,$this->userServerID);
+		list($lastName,$firstName,$lastNameEn,$firstNameEn,$pilotCountry,$Sex,$Birthdate,$CIVL_ID)=getPilotInfoEn($this->userID,$this->userServerID);
 
 		$userServerID=$this->userServerID;
 		if ($userServerID==0) $userServerID=$CONF_server_id;
@@ -333,6 +355,8 @@ var $maxPointNum=1000;
 	<userName>$this->userName</userName>
 	<pilotFirstName>$firstName</pilotFirstName>
 	<pilotLastName>$lastName</pilotLastName>
+	<pilotFirstNameEn>$firstNameEn</pilotFirstNameEn>
+	<pilotLastNameEn>$lastNameEn</pilotLastNameEn>
 	<pilotCountry>$pilotCountry</pilotCountry>
 	<pilotBirthdate>$Birthdate</pilotBirthdate>
 	<pilotSex>$Sex</pilotSex>
@@ -428,6 +452,8 @@ $resStr='{
 		"userName": "'.json::prepStr($lastName.' '.$firstName).'",
 		"pilotFirstName": "'.json::prepStr($firstName).'",
 		"pilotLastName": "'.json::prepStr($lastName).'",
+		"pilotFirstNameEn": "'.json::prepStr($firstNameEn).'",
+		"pilotLastNameEn": "'.json::prepStr($lastNameEn).'",
 		"pilotCountry": "'.$pilotCountry.'",
 		"pilotBirthdate": "'.json::prepStr($Birthdate).'",
 		"pilotSex": "'.$Sex.'"
@@ -551,8 +577,18 @@ $resStr='{
 
 	}
 
-	function checkBound($time){
+	function checkBound($time)
+	{
+		// GUS begin
+		if ( $this->START_TIME <= $this->END_TIME )
+		{
 		if ($time >= $this->START_TIME &&  $time <= $this->END_TIME ) return 1;
+		}
+		else
+		{
+			if ($time >= $this->START_TIME || $time <= $this->END_TIME) return 1;
+		}
+		// GUS end
 		return 0;
 	}
 
@@ -1551,12 +1587,12 @@ $kml_file_contents.="
 
     function createStaticMap($forceRefresh=0) {
 
-        $maxPoints=400;
-        $width=350;
-        $height=350;
+        $maxPoints=250;
+        $width=450;
+        $height=450;
 
         global $moduleRelPath,$baseInstallationPath;
-        global $langEncodings,$currentlang;
+        global $langEncodings,$currentlang,$CONF;
 
         if ( is_file($this->getMapFilename())  && !$forceRefresh )  return $this->getMapRelPath();
 
@@ -1570,8 +1606,12 @@ $kml_file_contents.="
         $prevLat=0;
         $prevLon=0;
 
+		$totLines=0;
+		foreach($lines as $line) {
+			if ($line{0}=='B' && strlen($line) >= 23 && $line{24}!='V' ) $totLines++;
+		}
 
-        $totLines=count($lines);
+        //$totLines=count($lines);
        // echo "tot line: ".$totLines;
 
         $mod0=0;
@@ -1634,18 +1674,53 @@ $kml_file_contents.="
 
         // echo "<br>encoded len: ".strlen ($kml_file_contents);
 
-        $staticMap="https://maps.googleapis.com/maps/api/staticmap?path=color:0xff0000|weight:2|enc:$kml_file_contents&maptype=terrain&size=".$width."x".$height."&sensor=false";
-        //echo "<br>".strlen($staticMap);
+        $flightScore=new flightScore($this->flightID);
+        $flightScore->getFromDB();
+
+        $defaultMethodID= $CONF['scoring']['default_set'];
+        $scoreDetails=$flightScore->scores[$defaultMethodID][ $flightScore->bestScoreType ];
+
+        //print_r($scoreDetails);
+        $j=0;
+        $markers="";
+        for($i=1;$i<=7;$i++) {
+            if ($scoreDetails['tp'][$i]) {
+                $newPoint=new gpsPoint($scoreDetails['tp'][$i],$this->timezone);
+                //$url.=urlencode('|'.sprintf("%.3f",$newPoint->lat).','.sprintf("%.3f",-$newPoint->lon));
+                if ($j==0) {
+                    $color='0x0000ff';
+                    $t='T';
+                } else {
+                    $color='0x00ff00';
+                    $t=$j;
+                }
+
+                if (!$scoreDetails['tp'][$i+1]) {
+                    $t = "L";
+                    $color='0xffff00';
+                }
+
+
+
+                $markers.="&markers=".urlencode("color:$color|label:$t|".sprintf("%.3f",$newPoint->lat).",".sprintf("%.3f",-$newPoint->lon) );
+                $j++;
+            }
+        }
+
+
+
+        $staticMap="https://maps.googleapis.com/maps/api/staticmap?path=color:0xff0000|weight:2|enc:$kml_file_contents".$markers."&maptype=terrain&size=".$width."x".$height."&sensor=false";
+        //echo "<br>".$staticMap;
 
         // get from google
         $imgStr=file_get_contents($staticMap);
-        echo "<br>".strlen($imgStr);
+        // echo "<br>".strlen($imgStr);
 
         $handle = fopen($this->getMapFilename(), "w");
         fwrite($handle, $imgStr);
         fclose($handle);
 
-        echo "<img src='".$this->getMapRelPath()."'>";
+        // echo "<img src='".$this->getMapRelPath()."'>";
 
         return $this->getMapRelPath();
     }
@@ -1795,9 +1870,10 @@ $kml_file_contents=
 					$thisPoint->gpsTime+=$day_offset;
 
 					// check for start of flight
-					if ($thisPoint->getTime()<$this->START_TIME) continue;
+				//if ($thisPoint->getTime() < $this->START_TIME) continue;
 					// check for end of flight
-					if ($thisPoint->getTime()>$this->END_TIME) continue;
+				//if ($thisPoint->getTime() > $this->END_TIME) continue;
+				if ( !$this->checkBound($thisPoint->getTime()) ) continue;
 
 					$data_time[$i]=sec2Time($thisPoint->getTime(),1);
 					$data_alt[$i]=$thisPoint->getAlt();
@@ -1819,6 +1895,7 @@ $kml_file_contents=
 						}
 						        		        	
 						$deltaseconds = $thisPoint->getTime() - $lastPoint->getTime() ;
+					if ( $deltaseconds < -85000 ) $deltaseconds += 86400;   // GUS in
 						if ($deltaseconds) {
 							$speed = ($deltaseconds)?$tmp*3.6/($deltaseconds):0.0; /* in km/h */
 							$data_vario[$i] =($thisPoint->getAlt() - $lastPoint->getAlt() ) / $deltaseconds;
@@ -1873,14 +1950,16 @@ $kml_file_contents=
 			$line = trim($lines[$k]);
 			if (strlen($line) == 0) continue;
 
+			//DEBUG("IGC", 1, "GUS PointsFilename [$k]: $line<br>");
+
 			eval($line);
 
 			//	if ($alt > $this->maxAllowedHeight)  continue;
 			//    if ($speed > $this->maxAllowedSpeed) continue;
 			//    if (abs($vario) > $this->maxAllowedVario) continue;
 
-			if  ( $time<$lastPointTime ) continue;
-			$lastPointTime=$time;
+			//if ($time < $lastPointTime) continue;
+			//$lastPointTime = $time;
 
 			//if (! $time_in_secs ) {
 			//	$time = sec2Time($time, 1);
@@ -1895,10 +1974,6 @@ $kml_file_contents=
 			$jsTrack['vario'][$i] 	= sprintf('%.2f',$vario);
 			
 			
-			
-
-
-
 			if ( $CONF['maps']['3d'] ) {
 				require_once dirname(__FILE__).'/CL_hgt.php';
 				$jsTrack['elevGnd'][$i] = $elevGnd[$i] = hgt::getHeight($lat,-$lon);
@@ -1970,6 +2045,7 @@ $kml_file_contents=
     $lines = file($this->getPointsFilename(1)); // get the normalized with constant time step points array
     if (!$lines) return;
     $i = 0;
+
 
 	// first 3 lines of pointsFile is reserved for info
 	for($k=3;$k< count($lines);$k++){
@@ -2090,8 +2166,18 @@ $kml_file_contents=
 		$currentDate=0;
 		$prevDate=0;
 
+		// GUS begin
+		$interval = 20;
+		$lastSnap = -1;
+		$normOrder = array();
+		// GUS end
+
+		$limiter = max(1, intval(count($lines) / 8000)); // GUS in for limiter
+		$li = 0;        // GUS in for limiter
+
 		foreach($lines as $line) {
 			$line=trim($line);
+
 			if  (strlen($line)==0) continue;
 
 			if (strtoupper(substr($line,0,5)) =="HFDTE"  || strtoupper(substr($line,0,5)) =="HPDTE"  ) {  // HFDTE170104  OR HPDTE310805
@@ -2112,10 +2198,13 @@ $kml_file_contents=
 
 			if ($line{0}=='B') {
 					if  ( strlen($line) < 23 ) 	continue;
+
 					// also check for bad points
 					// 012345678901234567890123456789
 					// B1522144902558N00848090EV0058400000
 					if ($line{24}=='V') continue;
+
+				if ( $li++ % $limiter ) continue; // GUS in for limiter
 
 					$thisPoint=new gpsPoint($line,$this->timezone);
 					$thisPoint->gpsTime+=$day_offset;
@@ -2129,14 +2218,13 @@ $kml_file_contents=
 
 					if ($i>0) {
 						$tmp = $lastPoint->calcDistance($thisPoint);
-
 					$time_diff=  $thisPoint->getTime() - $lastPoint->getTime()  ;
-
-					if (  $time_diff < 0 && $time_diff > -36000  )  { // if the time is less than 10 hours in the past  we just ignore it
+					if (( $time_diff < 0) && ( $time_diff > -36000 ) ) { // if the time is less than 10 hours in the past  we just ignore it
 						continue;
-        		    } else 	if ( $time_diff < 0  )  {  // CHANGING DAY , means the flight is at night
+					// GUS out:
+					/*} else if ($time_diff < 0) {  // CHANGING DAY , means the flight is at night
 						array_pop($goodPoints);
-						break;
+						break;*/
         		    } else 	if (  $time_diff > $this->max_allowed_time_gap *4 )  {  // found time gap
 						array_pop($goodPoints);
 						break;
@@ -2153,9 +2241,13 @@ $kml_file_contents=
 						*/
 
 						$deltaseconds = $thisPoint->getTime() - $lastPoint->getTime() ;
+
+					if ( $deltaseconds < -85000 ) $deltaseconds += 86400;   // GUS in
 						if ($deltaseconds) {
 							$speed = ($deltaseconds)?$tmp*3.6/($deltaseconds):0.0; /* in km/h */
+
 							$goodPoints[$i]['vario'] =($thisPoint->getAlt() - $lastPoint->getAlt() ) / $deltaseconds;
+
 						} else {
 							$speed =0;
 							$data_vario[$i]=0;
@@ -2169,8 +2261,20 @@ $kml_file_contents=
 						$takeoffPoint=new gpsPoint($line,$this->timezone);
 						$goodPoints[$i]['dis']=0;
 					}
+
 					$lastPoint=new gpsPoint($line,$this->timezone);
 					$lastPoint->gpsTime+=$day_offset;
+
+				// GUS begin
+				$timeSnap = intval($lastPoint->getTime() / $interval) * $interval;
+				if ( $lastSnap != $timeSnap )
+				{
+					$lastSnap = $timeSnap;
+					$normIndex[$lastSnap] = $goodPoints[$i];
+					$normOrder[] = $lastSnap;
+					//DEBUG('IGC',1, "[time: $i] $line<br>");
+				}
+				// GUS end
 
 					$i++;
 			}
@@ -2183,16 +2287,20 @@ $kml_file_contents=
 		$max_time=ceil($goodPoints[$pointsNum-1]['time']/60)*60;
 		// echo "max / min time: $max_time $min_time<BR>";
 
-		$interval=20;
 		$k=0;
-		for($i=$min_time;$i<=$max_time;$i=$i+$interval) {
-			$normPoints[$k]=array();
+		// GUS begin
+		$o = 0;
+		for ($i = $min_time; $i != $max_time; $i += $interval) {
+			if ( $i >= 86400 ) $i -= 86400;
+			$from = ( isset($normIndex[$i]) ) ? $o++ : $o ;
+			$normPoints[$k] = $normIndex[$normOrder[$from]];
 			$normPoints[$k]['timeText']=sec2time($i);
 			$normPoints[$k]['time']=$i;
 			$k++;
 		}
+		// GUS end
 
-		$org_array_pos=0;
+		/*$org_array_pos = 0;
 		foreach($normPoints as $norm_array_pos=>$normPoint) { // for each point  in the timeline
 			$timeval=$normPoint['time'];
 			if ( $goodPoints[$org_array_pos]['time'] >= $timeval && $org_array_pos< count($goodPoints)  &&
@@ -2211,7 +2319,7 @@ $kml_file_contents=
 				while ($goodPoints[$org_array_pos]['time'] <  $timeval + $interval  && ($org_array_pos)< count($goodPoints) ) {
 					$org_array_pos++;
 				}
-			} else {
+			} else if ( !isset($normPoints[$norm_array_pos]['lon']) ) {     // GUS in
 				if ($norm_array_pos ) {
 					$normPoints[$norm_array_pos]['lon']=$normPoints[$norm_array_pos-1]['lon'];
 					$normPoints[$norm_array_pos]['lat']=$normPoints[$norm_array_pos-1]['lat'];
@@ -2228,18 +2336,15 @@ $kml_file_contents=
 					$normPoints[$norm_array_pos]['alt']=$goodPoints[0]['alt'];
 					$normPoints[$norm_array_pos]['altV']=$goodPoints[0]['altV'];
 					$normPoints[$norm_array_pos]['dis']=$goodPoints[0]['dis'];
-				}
+				}*/
 	/*
 				if ($org_array_pos>0) {
 					$normPoints[$norm_array_pos]['speed']="-";
 					$normPoints[$norm_array_pos]['vario']="-";
 					$normPoints[$norm_array_pos]['alt']="-";
 				}
-	*/
-			}
-
-
-		} // end foreach
+	*/ /*             }
+		} // end foreach*/
 
 		// now write it to file
 		$outputBuffer='$min_time='.$min_time.';$max_time='.$max_time.";\n\n\n";
@@ -2441,9 +2546,8 @@ $kml_file_contents=
 
 	/****************************************************************************************
 	****************************************************************************************
-
-		Core function of analizing an igc
-
+	 *
+	 * Core function of analizing an igc
 	****************************************************************************************
 	****************************************************************************************/
 	function getFlightFromIGC($filename) {
@@ -2492,6 +2596,11 @@ $kml_file_contents=
 
 				if  (strlen($line)==0) continue;
 				if  ( $line{0}!='B' ) continue;
+				/* if ($i>2){
+					$prevline=trim($lines[$i-1]);
+					if  (!substr_compare($line,$prevline,1,6)) continue; 
+					//throw out Brecords with same time signatures P. Wild 4.11.2012
+				} */
 				$Brecords++;
 
 				// also check for bad points
@@ -2777,7 +2886,7 @@ $kml_file_contents=
 							DEBUG("IGC",1,"[$points] $line<br>");
 							DEBUG("IGC",1,"[$points] Point in the past<br>");
 							continue;
-        		    } else 	if ( $time_diff < 0  )  {  // CHANGING DAY , means the flight is at night
+					} else if (($time_diff < 0) && ($time_diff > -86000)) {  // CHANGING DAY , means the flight is at night
 						$lastPoint=$prevPoint;
 						$foundNewTrack=1;
 						DEBUG("IGC",1,"[$points] $line<br>");
@@ -2805,9 +2914,13 @@ $kml_file_contents=
 					// compute some things
 					$tmp = $lastPoint->calcDistance($prevPoint);
 					$alt = $lastPoint->getAlt();
+					// GUS begin
 					$deltaseconds = $lastPoint->getTime() - $prevPoint->getTime() ;
-					$speed = ($deltaseconds)?$tmp*3.6/($deltaseconds):0.0; /* in km/h */
-					if ($deltaseconds) $vario=($alt-$prevPoint->getAlt() ) / $deltaseconds;
+					$speedDeltaSecs = $deltaseconds + ( ( $deltaseconds < -85000 ) ? 86400 : 0);
+					$speed = ($speedDeltaSecs) ? $tmp * 3.6 / ($speedDeltaSecs) : 0.0; /* in km/h */
+					if ($speedDeltaSecs) $vario = ($alt - $prevPoint->getAlt()) / $speedDeltaSecs;
+					// GUS end
+
 
 					if (!$garminSpecialCase && ! $this->forceBounds) {
 						if ( ($fast_points>=$pointsNeededForTakeoff || $fast_points_dt>30) && $stillOnGround) { // found 5 flying points or 30 secs
@@ -2820,7 +2933,7 @@ $kml_file_contents=
 							// either speed >= 15 or if we already have 2 fast points settle with speed>=10
 							if ($speed >= 15 ||  ( $speed >= 10 && $fast_points>=2 ) ) {
 								$fast_points++;
-								$fast_points_dt+=$deltaseconds;
+								$fast_points_dt += $speedDeltaSecs;
 								DEBUG("IGC",1,"[$points] $line<br>");
 								DEBUG("IGC",1,"[$points] Found a fast speed point <br>");
 							} else { // reset takeoff scan
@@ -2835,9 +2948,9 @@ $kml_file_contents=
 						} else { //landing  scan
 							if ($speed < 5 ) {
 								$slow_points++;
-								$slow_points_dt+=$deltaseconds;
+								$slow_points_dt += $speedDeltaSecs;
 								DEBUG("IGC",1,"[$points] $line<br>");
-								DEBUG("IGC",1,"[$points] Found a slow speed point (speed, dt)=($speed,$deltaseconds)<br>");
+								DEBUG("IGC", 1, "[$points] Found a slow speed point (speed, dt)=($speed,$speedDeltaSecs)<br>");
 							} else {
 								$slow_points=0;
 								$slow_points_dt=0;
@@ -2866,8 +2979,11 @@ $kml_file_contents=
 						$tmp = $lastPoint->calcDistance($prevPoint2);
 						$alt = $lastPoint->getAlt();
 						$deltaseconds = $time_diff2;
-						$speed = ($deltaseconds)?$tmp*3.6/($deltaseconds):0.0; /* in km/h */
-						if ($deltaseconds) $vario=($alt-$prevPoint2->getAlt() ) / $deltaseconds;
+						// GUS begin
+						$speedDeltaSecs = $deltaseconds + ( ( $deltaseconds < -85000 ) ? 86400 : 0);
+						$speed = ($speedDeltaSecs) ? $tmp * 3.6 / ($speedDeltaSecs) : 0.0; /* in km/h */
+						if ($speedDeltaSecs) $vario = ($alt - $prevPoint2->getAlt()) / $speedDeltaSecs;
+						// GUS end
 						$prevPoint2=new gpsPoint($line,$this->timezone);
 						$prevPoint2->gpsTime+=$day_offset;
 						
@@ -2928,6 +3044,9 @@ $kml_file_contents=
 		if (! writeFile($this->getIGCFilename(2),$fullSanedFile) ) {
 		   echo "Problem writing to file (".$this->getIGCFilename(2).")";
 		}
+
+		DEBUG("IGC", 1, "<HR>". $this->getIGCFilename(2) . ' size: ' .strlen($fullSanedFile)."<HR>");
+
 		// echo "<HR><HR>". $this->getIGCFilename(2) .strlen($fullSanedFile)."<HR><HR>";
 		/* done wrting the full saned file */
 
@@ -2936,6 +3055,7 @@ $kml_file_contents=
 		   echo "Problem writing to file (".$this->getIGCFilename(1).")";
 		}
 		// done write saned IGC file
+		DEBUG("IGC", 1, "<HR>". $this->getIGCFilename(1) . ' size: ' .strlen($outputBuffer)."<HR>");
 
 		if ($lastPoint) {
 			$this->lastPointTM=$lastPoint->gpsTime;
@@ -2954,6 +3074,7 @@ $kml_file_contents=
 		}
 
 		$this->DURATION =   $this->END_TIME - $this->START_TIME ;
+		if ($this->DURATION < 0) $this->DURATION += 86400;
 		$this->MEAN_SPEED = $this->MEAN_SPEED / $MEAN_SPEED_POINTS;
 
 		return 1;
@@ -3557,6 +3678,7 @@ $kml_file_contents=
 
 		DEBUG("MAP",1,"marginHor: $marginHor, marginVert:$marginVert <br>");
 
+
 		$flMap=new flightMap($UTMzone,$UTMlatZone, $MAP_TOP + $marginVert, $MAP_LEFT - $marginHor,$UTMzone2, $UTMlatZone2, $MAP_BOTTOM - $marginVert ,$MAP_RIGHT +$marginHor  , 600,800,$this->getIGCFilename(1),$this->getMapFilename(0),$this->is3D() );
 		DEBUG("MAP",1,"MAP Required m/pixel = ".$flMap->metersPerPixel."<br>");
 		$flMap->drawFlightMap();
@@ -3582,8 +3704,8 @@ $kml_file_contents=
 		  }
 
 		  if (! $row = $db->sql_fetchrow($res) ) {
-		  echo "###### ERRROR ####$query###";
-			  return 0;
+		     // echo "###### ERRROR ####$query###";
+			 return 0;
 		  }
 	  }
 
@@ -3886,10 +4008,19 @@ $kml_file_contents=
 		global $flightsTable,$deletedFlightsTable;
 		global $CONF_photosPerFlight,$CONF_server_id;
 
-		$query="INSERT INTO $deletedFlightsTable SELECT $flightsTable.* from $flightsTable WHERE $flightsTable.ID=".$this->flightID." ";
+        if ($this->filename) {
+            // save a copy
+            $newfilename=$this->filename."_DELETED_".$this->flightID.".igc";
+            // echo "renaming igc file to back it up: $newfilename<br>";
+            $this->renameTracklog($newfilename);
+        }
+
+        $query="INSERT INTO $deletedFlightsTable SELECT $flightsTable.* from $flightsTable WHERE $flightsTable.ID=".$this->flightID." ";
+        //echo $query;
 		$res= $db->sql_query($query);
 
         $query="UPDATE $deletedFlightsTable Set dateUpdated='".gmdate("Y-m-d H:i:s")."' where ID=".$this->flightID." ";
+        //echo $query;
 	    $res= $db->sql_query($query);
 
 		$query="DELETE from $flightsTable  WHERE ID=".$this->flightID." ";
@@ -3901,14 +4032,16 @@ $kml_file_contents=
 		  exit();
  	    }
 
-		// save a copy
+
 		$this->flightScore=new flightScore($this->flightID);
 		$this->flightScore->getFromDB();
 
 		$flightScore=new flightScore($this->flightID);
 		$flightScore->deleteFromDB();
-		// Now delete the files
 
+
+		// Now delete the files
+if (1){
 		$this->deleteFile($this->getIGCFilename(0) );
 		$this->deleteFile($this->getIGCFilename(0).".olc" );
 		
@@ -3929,6 +4062,7 @@ $kml_file_contents=
 				$this->deleteFile($this->getChartFilename("takeoff_distance",$metric_system,$raw) );
 			}
 		}
+   }
 
 		if ($this->hasPhotos) {
 			$flightPhotos=new flightPhotos($this->flightID);
